+++ face_attributes_detection/MTL_no_gender_cv_best_model.py
@@ -1,152 +1,23 @@
 import os
 import sys
-import time
-import math
 import getopt
-import statistics
+from collections import defaultdict
+
 import numpy as np
-import cv2
 import pandas as pd
-import tensorflow as tf
 import matplotlib.pyplot as plt
-import keras.backend as K
-from tensorflow.keras.models import Model
-from tensorflow.keras.layers import Conv2D, Dropout, MaxPooling2D, Activation, Dense, Flatten, Input, \
-    BatchNormalization, Lambda
-from tensorflow.keras.utils import Sequence
-from sklearn.model_selection import KFold
-import platform
+
+from tensorflow.keras.callbacks import ReduceLROnPlateau, EarlyStopping, ModelCheckpoint
+from tensorflow import optimizers as topt
+
+from platform_settings import *
+from common import *
+from networks import FaceNet_NoGender
+from sequences import *
+from utils import *
+
 from contextlib import redirect_stdout
 
-host = platform.node()
-aligned = True
-root_linux = "/dev/shm/data/celeba_files/"
-root_windows = "C:/Users/Ismail/Documents/Projects/celeba_files/"
-root_scaleway = '/root/data/celeba_files/'
-
-if host == 'castor' or host == 'altair':  # Enrico's PCs
-    root_path = root_linux
-elif host == 'DESKTOP-AS5V6C3':  # Ismail's PC
-    root_path = root_windows
-elif host == 'scw-zealous-ramanujan':
-    root_path = root_scaleway
-else:
-    raise RuntimeError('Unknown host')
-
-global_path = root_path
-
-if aligned:
-    images_path = root_path + "well_cropped_images/"
-    attributes_path = root_path + "celeba_csv/list_attr_celeba_aligned.csv"
-else:
-    images_path = root_path + "cropped_images/"
-    attributes_path = root_path + "celeba_csv/list_attr_celeba.csv"
-
-def build_folder(path):
-    # build a directory and confirm execution with a message
-    try:
-        os.makedirs(path)
-        print("<== {} directory created ==>")
-    except OSError:
-        print("Creation of the directory %s failed" % path)
-
-def update(dict_col, att_dict, flag='initialize', compute_flops=True):
-    if flag == 'initialize':
-        dict_col["number of Conv2D"] = []
-        dict_col["number of Dense"] = []
-        dict_col["kernel size"] = []
-        dict_col["first conv"] = []
-        dict_col["second conv"] = []
-        dict_col["unit"] = []
-        dict_col["batch_size"] = []
-        for key, value in att_dict.items():
-            dict_col[key + " cv score"] = []
-            dict_col[key + " std score"] = []
-        if compute_flops == True:
-            dict_col["flop"] = []
-
-def multiple_append(listes, elements):
-    # append different elements to different lists in the same time
-    if len(listes) != len(elements):
-        # ensure, there is no mistake in the arguments
-        raise RuntimeError("lists and elements do not have the same length")
-    else:
-        for l, e in zip(listes, elements):
-            l.append(e)
-
-def labelize(outputs):
-    # transform the vector output of a final dense layer with softmax
-    # the most likely label gets one and the other takes 0
-    # as would .utils.to_categorical do to a binary categorical attributes
-    return np.argmax(outputs, axis=1)
-
-def pred_to_label(prediction, attribute):
-    if attribute == 'mustache':
-        if prediction == 0:
-            return 'no mustache'
-        else:
-            return 'mustache'
-
-    elif attribute == 'eyeglasses':
-        if prediction == 0:
-            return 'no eyeglasses'
-        else:
-            return 'eyeglasses'
-
-    elif attribute == 'beard':
-        if prediction == 0:
-            return 'no beard'
-        else:
-            return 'beard'
-
-    elif attribute == 'hat':
-        if prediction == 0:
-            return 'no hat'
-        else:
-            return 'wearing hat'
-
-    else:
-        if prediction == 0:
-            return 'hairy'
-        else:
-            return 'bald'
-
-def predict(model, test_images, flag='class'):
-    predictions, adapted_images = [], []
-    predicted_mustache, predicted_eyeglasses, \
-    predicted_beard, predicted_hat, predicted_bald = [], [], [], [], []
-
-    for image in test_images:
-        img = image.reshape(-1, 36, 36, 1)
-        prediction = model.predict(img)
-        mustache_predict, eyeglasses_predict, \
-        beard_predict, hat_predict, bald_predict = np.argmax(prediction, axis=2)
-        if flag == 'class':
-            multiple_append([predicted_mustache, predicted_eyeglasses,
-                             predicted_beard, predicted_hat, predicted_bald],
-                            [mustache_predict[0], eyeglasses_predict[0],
-                             beard_predict[0], hat_predict[0], bald_predict[0]])
-        elif flag == 'label':
-            multiple_append([predicted_mustache, predicted_eyeglasses,
-                             predicted_beard, predicted_hat, predicted_bald],
-                            [pred_to_label(mustache_predict[0], 'mustache'),
-                             pred_to_label(eyeglasses_predict[0], 'eyeglasses'),
-                             pred_to_label(beard_predict[0], 'beard'),
-                             pred_to_label(hat_predict[0], 'hat'),
-                             pred_to_label(bald_predict[0], 'bald')])
-
-    return predicted_mustache, predicted_eyeglasses, \
-           predicted_beard, predicted_hat, predicted_bald
-
-## TODO: it's a pity to import keras (which is already in tensorflow) just to not re-implement this
-def f1(y_true, y_pred):  # taken from old keras source code
-    true_positives = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
-    possible_positives = K.sum(K.round(K.clip(y_true, 0, 1)))
-    predicted_positives = K.sum(K.round(K.clip(y_pred, 0, 1)))
-    precision = true_positives / (predicted_positives + K.epsilon())
-    recall = true_positives / (possible_positives + K.epsilon())
-    f1_val = 2 * (precision * recall) / (precision + recall + K.epsilon())
-    return f1_val
 
 def plot_epoch(history, metric, filename):
     # takes history of model.fit and extract training and validation metric data
@@ -170,187 +41,11 @@ def plot_epoch(history, metric, filename):
     plt.savefig(filename)
     plt.close()
 
-def get_flops(model_h5_path):
-    # computes floating points operations for a h5 model
-    session = tf.compat.v1.Session()
-    graph = tf.compat.v1.get_default_graph()
-
-    with graph.as_default():
-        with session.as_default():
-            model = tf.keras.models.load_model(model_h5_path)
-            run_meta = tf.compat.v1.RunMetadata()
-            opts = tf.compat.v1.profiler.ProfileOptionBuilder.float_operation()
-            # We use the Keras session graph in the call to the profiler.
-            flops = tf.compat.v1.profiler.profile(graph=graph, run_meta=run_meta, cmd='op', options=opts)
-            return flops.total_float_ops
-
-def avg(liste):
-    if type(liste)!=list:
-        print("it is not a list")
-    else:
-        return sum(liste)/len(liste)
-
-def adapt(x):
-    return (x+1)/2
-
-def anti_adapt(x):
-    return (-x+1)/2
-
-class FaceNet:
-    def __init__(self, shape, channel, unit, first_conv, second_conv):
-        self.shape = shape
-        self.channel = channel
-        self.unit = unit
-        self.first = first_conv
-        self.second = second_conv
-        self.categs = ['mustache', 'eyeglasses', 'beard', 'hat', 'bald']
-
-    def build(self, size, num=2):
-        # initialize the input shape and channel dimension (this code
-        # assumes you are using TensorFlow which utilizes channels
-        # last ordering)
-        inputShape = (self.shape, self.shape, self.channel)
-
-        # construct mustache, bald, eyeglasses, beard & hat sub-networks
-        inputs = Input(shape=inputShape)
-        x = Conv2D(self.first, size, padding="same")(inputs)
-        x = BatchNormalization()(x)
-        x = Activation("relu")(x)
-        x = MaxPooling2D(pool_size=(2, 2))(x)
-        x = Conv2D(self.second, size, padding="same")(x)
-        x = BatchNormalization()(x)
-        x = Activation("relu")(x)
-        x = MaxPooling2D(pool_size=(2, 2))(x)
-
-        flat = Flatten()(x)
-        net = Dense(self.unit)(flat)
-        net = BatchNormalization(axis=-1)(net)
-        net = Activation("relu")(net)
-
-        outputs = []
-        for c in self.categs:
-            tmp_net = Dense(num)(net)
-            tmp_net = Activation('softmax', name=c)(tmp_net)
-            outputs.append(tmp_net)
-
-        # create the model using our input (the batch of images) and
-        # six separate outputs
-        model = Model(inputs=inputs, outputs=outputs, name="facenet")
-
-        # return the constructed network architecture
-        return model
-
-class CelebASequence(Sequence):
-    TEST_PROPORTION = 0.1
-
-    def __init__(self, attributes_path, images_path, batch_size, shape, channel, max_items=None, n_split=5):
-        self.images_path = images_path
-        self.batch_size = batch_size
-        self.sizes = (shape, shape, channel)
-        self.samples_per_data = 4
-        if batch_size % self.samples_per_data != 0:  # VERY debatable
-            raise NotImplementedError('Batch size has to be multiple of 4')
-
-        self.attributes_tab = pd.read_csv(attributes_path)
-        if max_items is not None:
-            self.attributes_tab = self.attributes_tab.iloc[0:max_items]
-
-        num_elements = len(self.attributes_tab['image_id'])
-        print('Full Dataset has ' + str(num_elements) + ' elements (before augmentation)')
-
-        indexes = np.arange(0, num_elements)
-
-        kf = KFold(n_splits=n_split,
-                   shuffle=True,
-                   random_state=42)
-
-        self.input_train, self.input_test = [], []
-        for train_index, test_index in kf.split(indexes):
-            multiple_append([self.input_train, self.input_test], [train_index, test_index])
-
-        print(f'After 5-Fold split: {len(self.input_train)} train and {len(self.input_test)} test')
-        self.set_mode_train()
-        self.set_mode_fold(0)
-        self.inner_batch_size = self.batch_size / self.samples_per_data
-
-        self.attributes = ['Mustache', 'Eyeglasses', 'No_Beard', 'Wearing_Hat', 'Bald']
-        self.attr_mapper = {'Mustache': 'mustache', 'Eyeglasses': 'eyeglasses', 'No_Beard': 'beard',
-                            'Wearing_Hat': 'hat', 'Bald': 'bald'}
-
-    def set_mode_train(self):
-        self.mode = 0
-
-    def set_mode_test(self):
-        self.mode = 1
-
-    def set_mode_fold(self, num_fold):
-        self.fold = num_fold
-
-    def __len__(self):
-        if self.mode == 0:
-            ln = len(self.input_train[self.fold])
-        else:
-            ln = len(self.input_test[self.fold])
-        return math.floor((ln * self.samples_per_data) / self.batch_size)
-
-    def __getitem__(self, idx):
-        st, sp = int(idx * self.inner_batch_size), int((idx + 1) * self.inner_batch_size)
-
-        imgs = np.empty((self.batch_size, *self.sizes))
-        atts = {'mustache': [], 'eyeglasses': [], 'beard': [], 'hat': [], 'bald': []}
-        j = 0
-
-        for k in range(st, sp):
-            if self.mode == 0:
-                index = self.input_train[self.fold][k]
-            else:
-                index = self.input_test[self.fold][k]
-
-            image_name = self.attributes_tab['image_id'][index]
-            img = cv2.cvtColor(cv2.imread(self.images_path + image_name), cv2.COLOR_BGR2GRAY) / 255
-            img = img.reshape(self.sizes)
-            img_b = cv2.blur(img, (2, 2)).reshape(self.sizes)
-            img_m = cv2.flip(img, 1).reshape(self.sizes)
-            img_mb = cv2.flip(img_b, 1).reshape(self.sizes)
-
-            imgs[j, :, :, :] = img
-            imgs[j + 1, :, :, :] = img_b
-            imgs[j + 2, :, :, :] = img_m
-            imgs[j + 3, :, :, :] = img_mb
-            j += self.samples_per_data
-
-            for a in self.attributes:
-                name = self.attr_mapper[a]
-                for b in range(0, self.samples_per_data):
-                    if name != 'beard':
-                        atts[name].append(adapt(self.attributes_tab[a][index]))
-                    else:
-                        atts[name].append(anti_adapt(self.attributes_tab[a][index]))
-
-        out_attrs = {}
-        for k, v in atts.items():
-            out_attrs[k] = tf.keras.utils.to_categorical(v, num_classes=2)
-
-        return (imgs, out_attrs)
-
-    def get_results(self):  # unused method
-        if self.mode != 1:
-            raise RuntimeError('Not in test mode')
-        atts = {'mustache': [], 'eyeglasses': [], 'beard': [], 'hat': [], 'bald': []}
-        for index in self.input_test:
-            for a in self.attributes:
-                name = self.attr_mapper[a]
-                for b in range(0, 4):
-                    if name != 'beard':
-                        atts[name].append(adapt(self.attributes_tab[a][index]))
-                    else:
-                        atts[name].append(anti_adapt(self.attributes_tab[a][index]))
 
 unit, first_conv, second_conv, batch_size, k_size = 16, 4, 16, 64, (3,3)
 
-def main(epochs=25, max_items=None):
+def main(epochs, max_items):
     shape, channel, compute_flops = 36, 1, True
-    s = 0
     losses = {"mustache": "categorical_crossentropy",
               "eyeglasses": "categorical_crossentropy",
               "beard": "categorical_crossentropy",
@@ -361,23 +56,27 @@ def main(epochs=25, max_items=None):
 
     # fit labels with images with a ReduceLRonPlateau which divide learning by 10
     # if for 2 consecutive epochs, global validation loss doesn't decrease
-    reducelronplateau = tf.keras.callbacks.ReduceLROnPlateau(monitor='loss', mode='min', patience=2, factor=0.1)
-    earlystopping = tf.keras.callbacks.EarlyStopping(monitor='loss', mode='min', patience=3)
+    reducelronplateau = ReduceLROnPlateau(monitor='loss', mode='min', patience=2, factor=0.1)
+    earlystopping = EarlyStopping(monitor='loss', mode='min', patience=3)
 
     # to have the flops we have to do that with a h5 model
     # problem is that you can't compile model with a f1 measure as it doesn't exist in keras originally
     # so, i retrain the model in one epoch, save it and then, compute flops of the model
     model_filename = root_path + "facenet_flops_test.h5"
-    checkpoint = tf.keras.callbacks.ModelCheckpoint(filepath=model_filename, monitor='loss', mode='min')
+    checkpoint = ModelCheckpoint(filepath=model_filename, monitor='loss', mode='min')
 
-    opt = tf.optimizers.SGD(lr=0.001)
+    opt = topt.SGD(lr=0.001)
 
-    seq = CelebASequence(attributes_path, images_path, batch_size, shape, channel, max_items=max_items)
-    s+=1
-    print(f"combinaison: {s}")
+    seq = CelebASequence(attributes_path, images_path, shape, channel, max_items=max_items)
+    seq.augment(Mirroring())
+    seq.augment(Blurring())
+    seq.augment(Blurring(Mirroring()))
+    seq.augment(MotionBlur('H'))
+    seq.augment(MotionBlur('H', Mirroring()))
+    seq.prepare(batch_size)
 
     #Creating the net for all these parameters
-    net = FaceNet(shape, channel, unit, first_conv, second_conv)
+    net = FaceNet_NoGender(shape, channel, unit, first_conv, second_conv)
     model = net.build(k_size)
 
     with open(root_path + 'modelsummary_no_gender.txt', 'w') as f:
@@ -388,9 +87,8 @@ def main(epochs=25, max_items=None):
     print("[INFO] compiling model...")
     model.compile(optimizer=opt, loss=losses, loss_weights=lossWeights, metrics=[f1, 'accuracy'])
 
-    bald_list, beard_list, hat_list, mustache_list, eyeglasses_list = [], [], [], [], []
-
     #Cross Validation 5-Fold
+    f1_scores = defaultdict(list)
     for k in range(5):
         #Train on 80%
         seq.set_mode_fold(k)
@@ -401,51 +99,46 @@ def main(epochs=25, max_items=None):
         #Test on 20%
         seq.set_mode_test()
         evaluations = model.evaluate(seq)
-
         print(f"evaluations are: {evaluations}")
 
-        bald_f1 = evaluations[-2]
-        hat_f1 = evaluations[-4]
-        beard_f1 = evaluations[-6]
-        eyeglasses_f1 = evaluations[-8]
-        mustache_f1 = evaluations[-10]
+        # read evaluations by indexes found in model.metrics_names
+        for kk in losses.keys():
+            idx = model.metrics_names.index(kk + '_f1')
+            f1_scores[kk].append(evaluations[idx])
 
-        multiple_append([bald_list, beard_list, hat_list, mustache_list, eyeglasses_list],
-                        [bald_f1, beard_f1, hat_f1, mustache_f1, eyeglasses_f1])
+    scores = {}
+    for kk in losses.keys():
+        score_array = np.array(f1_scores[kk], dtype='float64')
+        average = np.mean(score_array)
+        stddev = np.std(score_array)
+        scores[kk] = (average, stddev)
 
-    attributes_list = [mustache_list, eyeglasses_list, beard_list, hat_list, bald_list]
-    score_cv, score_std = [], []
-
-    for score_list in attributes_list:
-        score_cv.append(avg(score_list))
-        score_std.append(statistics.pstdev(np.array(score_list, dtype='float64')))
-
-    #compute flop
-    seq_fold = CelebASequence(attributes_path, images_path, batch_size, shape, channel,
-                         max_items=100)
-
+    ## Compute flop
+    # new sequence
+    seq_fold = CelebASequence(attributes_path, images_path, shape, channel, max_items=100)
+    seq_fold.prepare(batch_size)
     # initialize the optimizer and compile the model
     print("[INFO] compiling flop model...")
     model.compile(optimizer=opt, loss=losses, loss_weights=lossWeights, metrics=['accuracy'])
     seq_fold.set_mode_fold(0)
-    model.fit(x=seq, epochs=1, callbacks=[checkpoint])
-
-    #writing new file with performance and flop
+    model.fit(x=seq_fold, epochs=1, callbacks=[checkpoint])
     flop = get_flops(model_filename)
 
+    # writing new file with performance and flop
     file = open(root_path + "performance_cv_no_gender.txt", "w+")
-    i = 0
     for key, value in losses.items():
-        file.write(key + f" average score: {score_cv[i]}\n")
-        file.write(key + f" std score: {score_std[i]}\n")
-        i += 1
+        file.write(key + f" average score: {scores[key][0]}\n")
+        file.write(key + f" std score: {scores[key][1]}\n")
+
     file.write(f"model flop: {flop}")
 
-def usage():
+
+def usage(epochs, max_items):
     print('./' + os.path.basename(__file__) + ' [options]')
-    print('-e / --epochs N       Run training on N epochs [10]')
-    print('-n / --num_items N    Use at most N items from the dataset [all]')
+    print(f'-e / --epochs N       Run training on N epochs [{epochs}]')
+    print(f'-n / --num_items N    Use at most N items from the dataset [{"all" if max_items is None else str(max_items)}]')
     sys.exit(-1)
+
 
 if __name__ == '__main__':
     opts, args = getopt.getopt(sys.argv[1:], 'e:n:', ['epochs=', 'num_items='])
@@ -458,7 +151,7 @@ if __name__ == '__main__':
         elif o in ('-n', '--num_items'):
             max_items = int(a)
         else:
-            usage()
+            usage(epochs, max_items)
 
     main(epochs, max_items)
     sys.exit(0)
