+++ face_attributes_detection/MTL_no_gender_detection.py
@@ -1,94 +1,19 @@
 import os
 import sys
-import time
-import math
 import getopt
 from collections import defaultdict
 
 import numpy as np
-import cv2
 import pandas as pd
-import tensorflow as tf
-import matplotlib.pyplot as plt
-import keras.backend as K
 
-from tensorflow.keras.utils import Sequence
 from tensorflow.keras.callbacks import ReduceLROnPlateau, EarlyStopping, ModelCheckpoint
-from sklearn.model_selection import KFold
+from tensorflow import optimizers as topt
 
+from platform_settings import *
+from common import *
 from networks import FaceNet_NoGender
 from sequences import *
-from platform_settings import *
 from utils import *
-
-
-def initialize_results(att_dict, compute_flops=True):
-    dict_col = {}
-    dict_col["number of Conv2D"] = []
-    dict_col["number of Dense"] = []
-    dict_col["kernel size"] = []
-    dict_col["first conv"] = []
-    dict_col["second conv"] = []
-    dict_col["unit"] = []
-    dict_col["batch_size"] = []
-    for key, value in att_dict.items():
-        dict_col[key + " cv score"] = []
-        dict_col[key + " std score"] = []
-    if compute_flops == True:
-        dict_col["flop"] = []
-    return dict_col
-
-
-def predict(model, test_images, flag='class'):
-    predictions, adapted_images = [], []
-    predicted_mustache, predicted_eyeglasses, \
-    predicted_beard, predicted_hat, predicted_bald = [], [], [], [], []
-
-    for image in test_images:
-        img = image.reshape(-1, 36, 36, 1)
-        prediction = model.predict(img)
-        mustache_predict, eyeglasses_predict, \
-        beard_predict, hat_predict, bald_predict = np.argmax(prediction, axis=2)
-        if flag == 'class':
-            multiple_append([predicted_mustache, predicted_eyeglasses,
-                             predicted_beard, predicted_hat, predicted_bald],
-                            [mustache_predict[0], eyeglasses_predict[0],
-                             beard_predict[0], hat_predict[0], bald_predict[0]])
-        elif flag == 'label':
-            multiple_append([predicted_mustache, predicted_eyeglasses,
-                             predicted_beard, predicted_hat, predicted_bald],
-                            [pred_to_label(mustache_predict[0], 'mustache'),
-                             pred_to_label(eyeglasses_predict[0], 'eyeglasses'),
-                             pred_to_label(beard_predict[0], 'beard'),
-                             pred_to_label(hat_predict[0], 'hat'),
-                             pred_to_label(bald_predict[0], 'bald')])
-
-    return predicted_mustache, predicted_eyeglasses, \
-           predicted_beard, predicted_hat, predicted_bald
-
-## TODO: it's a pity to import keras (which is already in tensorflow) just to not re-implement this
-def f1(y_true, y_pred):  # taken from old keras source code
-    true_positives = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
-    possible_positives = K.sum(K.round(K.clip(y_true, 0, 1)))
-    predicted_positives = K.sum(K.round(K.clip(y_pred, 0, 1)))
-    precision = true_positives / (predicted_positives + K.epsilon())
-    recall = true_positives / (possible_positives + K.epsilon())
-    f1_val = 2 * (precision * recall) / (precision + recall + K.epsilon())
-    return f1_val
-
-def get_flops(model_h5_path):
-    # computes floating points operations for a h5 model
-    session = tf.compat.v1.Session()
-    graph = tf.compat.v1.get_default_graph()
-
-    with graph.as_default():
-        with session.as_default():
-            model = tf.keras.models.load_model(model_h5_path)
-            run_meta = tf.compat.v1.RunMetadata()
-            opts = tf.compat.v1.profiler.ProfileOptionBuilder.float_operation()
-            # We use the Keras session graph in the call to the profiler.
-            flops = tf.compat.v1.profiler.profile(graph=graph, run_meta=run_meta, cmd='op', options=opts)
-            return flops.total_float_ops
 
 
 k_sizes = [(3,3)]
@@ -121,7 +46,7 @@ def main(epochs, max_items, folds, skip):
         model_filename = root_path + "facenet_flops_test.h5"
         checkpoint = ModelCheckpoint(filepath=model_filename, monitor='loss', mode='min')
 
-    opt = tf.optimizers.SGD(lr=0.001)
+    opt = topt.SGD(lr=0.001)
 
     # build Sequence with data augmentation
     seq = CelebASequence(attributes_path, images_path, shape, channel, max_items=max_items)
@@ -134,7 +59,7 @@ def main(epochs, max_items, folds, skip):
     s = 0
     for k_size in k_sizes:
         for batch_size in batch_sizes:
-
+            # (Re)Initialize sequence on new batch size
             seq.prepare(batch_size)
 
             for first_conv in first_convs:
@@ -190,7 +115,7 @@ def main(epochs, max_items, folds, skip):
 
                             seq_flops = CelebASequence(attributes_path, images_path, shape, channel, max_items=100)
                             seq_flops.prepare(batch_size)
-                            opt_flops = tf.optimizers.SGD(lr=0.001)
+                            opt_flops = topt.SGD(lr=0.001)
 
                             print("[INFO] compiling flops model...")
                             model_flops.compile(optimizer=opt_flops, loss=losses, loss_weights=lossWeights, metrics=['accuracy'])
