+++ face_attributes_detection/utils.py
@@ -1,6 +1,8 @@
 
 import os
 import numpy as np
+import keras.backend as K
+import tensorflow as tf
 
 def multiple_append(listes, elements):
     # append different elements to different lists in the same time
@@ -18,6 +20,7 @@ def adapt(x):
 
 def anti_adapt(x):
     return (-x+1)/2
+
 
 def avg(liste):
     if type(liste)!=list:
@@ -42,33 +45,28 @@ def labelize(outputs):
     return np.argmax(outputs, axis=1)
 
 
-def pred_to_label(prediction, attribute):
-    if attribute == 'mustache':
-        if prediction == 0:
-            return 'no mustache'
-        else:
-            return 'mustache'
+def get_flops(model_h5_path):
+    # computes floating points operations for a h5 model
+    session = tf.compat.v1.Session()
+    graph = tf.compat.v1.get_default_graph()
 
-    elif attribute == 'eyeglasses':
-        if prediction == 0:
-            return 'no eyeglasses'
-        else:
-            return 'eyeglasses'
+    with graph.as_default():
+        with session.as_default():
+            model = tf.keras.models.load_model(model_h5_path)
+            run_meta = tf.compat.v1.RunMetadata()
+            opts = tf.compat.v1.profiler.ProfileOptionBuilder.float_operation()
+            # We use the Keras session graph in the call to the profiler.
+            flops = tf.compat.v1.profiler.profile(graph=graph, run_meta=run_meta, cmd='op', options=opts)
+            return flops.total_float_ops
 
-    elif attribute == 'beard':
-        if prediction == 0:
-            return 'no beard'
-        else:
-            return 'beard'
 
-    elif attribute == 'hat':
-        if prediction == 0:
-            return 'no hat'
-        else:
-            return 'wearing hat'
+## TODO: it's a pity to import keras (which is already in tensorflow) just to not re-implement this
+def f1(y_true, y_pred):  # taken from old keras source code
+    true_positives = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
+    possible_positives = K.sum(K.round(K.clip(y_true, 0, 1)))
+    predicted_positives = K.sum(K.round(K.clip(y_pred, 0, 1)))
+    precision = true_positives / (predicted_positives + K.epsilon())
+    recall = true_positives / (possible_positives + K.epsilon())
+    f1_val = 2 * (precision * recall) / (precision + recall + K.epsilon())
+    return f1_val
 
-    else:
-        if prediction == 0:
-            return 'hairy'
-        else:
-            return 'bald'
